LPS-134354 Move translation components to `frontend-taglib-clay`
