--- modules/apps/frontend-js/frontend-js-translation-web/.npmbridgerc
@@ -1,15 +1,0 @@
-{
-	"frontend-js-translation-web": {
-		"dest-mod-name-mapper": {
-			"from": "(.*)\\.js$",
-			"to": "bridge/@liferay/frontend-js-translation-web/$1"
-		},
-		"file-globs": "**/*.js,!**/bridge",
-		"input": "build/node/packageRunBuild/resources",
-		"output": "build/node/packageRunBuild/resources/bridge/@liferay/frontend-js-translation-web",
-		"src-mod-name-mapper": {
-			"from": "(.*)\\.js$",
-			"to": "@liferay/frontend-js-translation-web/$1"
-		}
-	}
-}
