--- modules/apps/frontend-js/frontend-js-translation-web/bnd.bnd
@@ -1,4 +1,0 @@
-Bundle-Name: Liferay Frontend JS Translation Web
-Bundle-SymbolicName: com.liferay.frontend.js.translation.web
-Bundle-Version: 1.0.0
-Web-ContextPath: /frontend-js-translation-web
