--- modules/apps/frontend-js/frontend-js-translation-web/build.gradle
@@ -1,8 +1,0 @@
-dependencies {
-	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "default"
-	compileOnly group: "org.apache.felix", name: "org.apache.felix.http.servlet-api", version: "1.1.2"
-	compileOnly group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.3.0"
-	compileOnly group: "org.osgi", name: "osgi.core", version: "6.0.0"
-	compileOnly project(":apps:frontend-js:frontend-js-loader-modules-extender-api")
-	compileOnly project(":core:petra:petra-string")
-}
