--- modules/apps/frontend-js/frontend-js-translation-web/package.json
@@ -1,26 +1,0 @@
-{
-	"dependencies": {
-		"@clayui/button": "3.6.0",
-		"@clayui/drop-down": "3.30.0",
-		"@clayui/form": "3.29.0",
-		"@clayui/icon": "3.1.0",
-		"@clayui/label": "3.4.1",
-		"@clayui/layout": "3.3.0",
-		"@clayui/management-toolbar": "3.3.0",
-		"@clayui/modal": "3.29.0",
-		"@clayui/table": "3.29.0",
-		"@liferay/frontend-js-react-web": "*",
-		"prop-types": "15.7.2",
-		"react": "16.12.0",
-		"react-dom": "16.12.0"
-	},
-	"main": "index.js",
-	"name": "@liferay/frontend-js-translation-web",
-	"scripts": {
-		"build": "liferay-npm-scripts build",
-		"checkFormat": "liferay-npm-scripts check",
-		"format": "liferay-npm-scripts fix",
-		"test": "liferay-npm-scripts test"
-	},
-	"version": "1.0.0"
-}
