--- modules/apps/frontend-js/frontend-js-translation-web/src/main/resources/content/Language.properties
@@ -1,3 +1,0 @@
-manage-translations=Manage Translations
-management-toolbar=Management Toolbar
-untranslated=Not Translated
