+++ modules/apps/frontend-taglib/frontend-taglib-clay/package.json
@@ -82,11 +82,13 @@
 		"react-dom": "16.12.0"
 	},
 	"name": "frontend-taglib-clay",
+	"main": "index.js",
 	"private": true,
 	"scripts": {
 		"build": "liferay-npm-scripts build",
 		"checkFormat": "liferay-npm-scripts check",
-		"format": "liferay-npm-scripts fix"
+		"format": "liferay-npm-scripts fix",
+		"test": "liferay-npm-scripts test"
 	},
 	"version": "7.1.15"
 }
