+++ modules/apps/frontend-taglib/frontend-taglib-clay/src/main/resources/META-INF/resources/translation/TranslationModal.js
