+++ modules/apps/frontend-taglib/frontend-taglib-clay/src/main/resources/content/Language.properties
@@ -7,16 +7,19 @@ final-price=Final Price
 inline-edit=Inline Edit
 item-was-successfully-created=Item was successfully created.
 item-was-successfully-updated=Item was successfully updated.
+manage-translations=Manage Translations
+management-toolbar=Management Toolbar
 no-fields-were-found=No fields were found.
-no-filters=No Filters
 no-filters-were-found=No filters were found.
-no-items=No Items
+no-filters=No Filters
 no-items-were-found=No items were found.
+no-items=No Items
 price-summary=Price Summary
 remove-x=Remove {0}
 reset-filters=Reset Filters
 set-filters=Set Filters
 show-hide-columns=Show/Hide columns
-table-data=Table Data
 table-data-updated=Table Data Updated
+table-data=Table Data
 unexpected-error=Unexpected Error
+untranslated=Not Translated
